// apps/humayro/lib/api/model/branch.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */
import type { ScoreBreakdown } from './scoreBreakdown';

export interface Branch {
  branchId?: number;
  branchName?: string;
  revenue?: number;
  orders?: number;
  averageCheck?: number;
  averageRating?: number;
  score?: number;
  scoreBreakdown?: ScoreBreakdown;
}